add about page content
