+++ src/app.css
@@ -12,7 +12,7 @@
 
 /* body full screen height */
 body {
-	height: 100vh;
+	min-height: 100vh;
 }
 
 /* body no scrollbar */
